pyproject.toml
[project]
name = "repl-nix-wash-and-go-delivery"
version = "0.1.0"
description = "A laundry service booking application using Streamlit with core features for scheduling pickups and managing orders"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "pillow>=10.4.0",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.9",
    "qrcode>=8.0",
    "requests>=2.32.3",
    "streamlit-webrtc>=0.47.9",
    "streamlit>=1.39.0",
    "stripe>=11.1.0",
]
